WIP
WIP

+++ src/main/java/com/flansmod/apocalypse/common/CommonProxyApocalypse.java
@@ -165,7 +165,7 @@ public class CommonProxyApocalypse
 						tags.setString("Type", type.shortName);
 						for(EnumDriveablePart part : EnumDriveablePart.values())
 						{
-							tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
+							tags.setFloat(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
 							tags.setBoolean(part.getShortName() + "_Fire", false);
 						}
 						DriveableData data = new DriveableData(tags);
