+++ src/main/java/com/flansmod/apocalypse/common/world/buildings/WorldGenResearchLab.java
@@ -284,7 +284,7 @@ public class WorldGenResearchLab extends WorldGenFlan
 		tags.setString("Type", type.shortName);
 		for(EnumDriveablePart part : EnumDriveablePart.values())
 		{
-			tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : type.health.get(part).health);
+			tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : (int) type.health.get(part).health);
 			tags.setBoolean(part.getShortName() + "_Fire", false);
 		}
 		for(int k = 0; k < 2; k++)
