+++ src/main/java/com/flansmod/apocalypse/common/world/buildings/WorldGenRunway.java
@@ -104,7 +104,7 @@ public class WorldGenRunway extends WorldGenFlan
 			tags.setString("Type", type.shortName);
 			for(EnumDriveablePart part : EnumDriveablePart.values())
 			{
-				tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : rand.nextInt(type.health.get(part).health));
+				tags.setInteger(part.getShortName() + "_Health", type.health.get(part) == null ? 0 : rand.nextInt((int) type.health.get(part).health));
 				tags.setBoolean(part.getShortName() + "_Fire", false);
 			}
